JobShop objects code cleaned with the use of new supplementary methods. Also, a bug in main script fixed. Old version expected only integer processing times for jobs. The test results for topologies 18 to 20 had to be fixed too (e.g. instead of 1 it is 1.0 in the result now)

+++ dream/simulation/LineGenerationJSON.py
@@ -306,7 +306,7 @@ def createObjects():
                 nextId=routeElement.get('stationId', 'not found')
                 processingTime=routeElement.get('processingTime', 'not found')
                 distributionType=processingTime.get('distributionType', 'not found')
-                mean=int(processingTime.get('mean', 'not found'))
+                mean=float(processingTime.get('mean', 'not found'))
                 route[stepNumber]=[nextId, mean]
             J=Job(id, name, route, priority=priority, dueDate=dueDate, orderDate=orderDate)
             G.JobList.append(J)
